+++ lib/features/services/screens/items.dart
@@ -5,6 +5,7 @@ import 'package:smart_city/features/services/controller.dart';
 import 'package:smart_city/features/services/screens/details.dart';
 import 'package:smart_city/features/services/screens/upsert.dart';
 import 'package:smart_city/features/services/widgets/company_tile.dart';
+import 'package:smart_city/shared/config/permissions.dart';
 import 'package:smart_city/shared/widgets/delete_confirm.dart';
 
 class ServicesCompaniesScreen extends StatelessWidget {
@@ -23,28 +24,32 @@ class ServicesCompaniesScreen extends StatelessWidget {
         centerTitle: true,
         title: Text(category['title_ru']),
         actions: [
-          IconButton(
-              onPressed: () async {
-                await showDialog(
-                    context: context,
-                    builder: (context) => DeleteConfirmAlert(
-                          onDelete: () async {
-                            await servicesController
-                                .deleteCategory(categoryId!);
-                          },
-                        ));
+          if (Permissions().getForServiceCategories()) ...[
+            IconButton(
+                onPressed: () async {
+                  await showDialog(
+                      context: context,
+                      builder: (context) => DeleteConfirmAlert(
+                            onDelete: () async {
+                              await servicesController
+                                  .deleteCategory(categoryId!);
+                            },
+                          ));
+                },
+                icon: const Icon(Icons.delete_rounded))
+          ],
+          if (Permissions().getForServices()) ...[
+            IconButton(
+              onPressed: () {
+                context.pushNamed(ServiceUpsert().route,
+                    queryParameters: {'categoryId': categoryId});
               },
-              icon: const Icon(Icons.delete_rounded)),
-          IconButton(
-            onPressed: () {
-              context.pushNamed(ServiceUpsert().route,
-                  queryParameters: {'categoryId': categoryId});
-            },
-            icon: const Icon(
-              Icons.add_rounded,
-              size: 30,
-            ),
-          )
+              icon: const Icon(
+                Icons.add_rounded,
+                size: 30,
+              ),
+            )
+          ]
         ],
       ),
       body: GridView.count(
